[controller] make ConfirmInstanceSyncHandler subclass SecureSyncHandler
- SecureSyncHandler subclasses SyncHandler, so the only functional difference
  is what this class adds
- SecureSyncHandler checks whether we are secure (using https) during prepare()
  and permanently redirects to use https and the subdomain "secure" on the same
  path and query if we are not.
- it is imperative to wait until after initialize() but before
  get/post/put/delete() to do this (that was kind of a bitch to figure out)

TODO: we really need a generic and simple URL class for generating URLs to our
own site uniformly. ideally it would make for simple construction of external
URLs as well, though. also, when we do that, internal URLs and regular
expressions referenced in urls.py should be pulled from constants for reuse.